+++ Packages/com.seahyx.xruitoolkit/Runtime/Core/Visual Effects/DropdownMenuNameAttribute.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace XRUIToolkit.Core.VisualEffect
+{
+	[AttributeUsage(AttributeTargets.Class)]
+	public class DropdownMenuNameAttribute : Attribute
+	{
+		public string MenuName { get; set; }
+
+		public DropdownMenuNameAttribute(string dropdownName)
+		{
+			MenuName = dropdownName;
+		}
+	}
+}