+++ Packages/com.seahyx.xruitoolkit/Runtime/Core/Visual Effects/Visual Effect Runtimes/AnimationTriggerVisualEffect.cs
@@ -2,7 +2,8 @@ using UnityEngine;
 
 namespace XRUIToolkit.Core.VisualEffect
 {
-	[CreateAssetMenu(fileName = "AnimationTriggerVFX", menuName = "HHUI Toolkit/Visual Effects/Animation Trigger VFX", order = 0)]
+	[CreateAssetMenu(fileName = "AnimationTriggerVFX", menuName = "HHUI Toolkit/Visual Effects/General/Animation Trigger VFX", order = 0)]
+	[DropdownMenuName("General/Animation Trigger VFX")]
 	public class AnimationTriggerVisualEffect : BaseVisualEffect
 	{
 		[SerializeField,