+++ Packages/com.seahyx.xruitoolkit/Runtime/Core/Visual Effects/Visual Effect Runtimes/MaterialColourVisualEffect.cs
@@ -3,7 +3,8 @@ using UnityEngine;
 
 namespace XRUIToolkit.Core.VisualEffect
 {
-	[CreateAssetMenu(fileName = "MaterialColourVFX", menuName = "HHUI Toolkit/Visual Effects/Material Colour VFX", order = 1)]
+	[CreateAssetMenu(fileName = "MaterialColourVFX", menuName = "HHUI Toolkit/Visual Effects/Colour/Material Colour VFX", order = 1)]
+	[DropdownMenuName("Colour/Material Colour VFX")]
 	public class MaterialColourVisualEffect : BaseAnimatedVisualEffect
 	{
 		protected enum ColourBlendMode