+++ Packages/com.seahyx.xruitoolkit/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.5.1
+
+### Changed
+
+- Changed "Create New Visual Effect" dropdown to use `GenericMenu` implementation
+- Updated the visual effect runtimes to be grouped by type
+- Updated the filename of visual effect runtimes to be consistent when created via dropdown or asset creation menu
+
 ## 1.5.0
 
 ### Added